carrito y libreria

+++ js/carrito.js
@@ -70,29 +70,86 @@ function actualizarBotonesEliminar() {
 }
 
 function eliminarDelCarrito(e) {
+  Swal.fire({
+    title: 'Estas seguro de querer eliminar este producto del carrito?',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Si, quiero eliminarlo',
+    cancelButtonText: 'No'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        'Eliminado!',
+        'El producto se elimino correctamente del carrito.',
+        'success'
+      )
+    }
+  });
   const idBoton = e.currentTarget.id;
   const index = productosEnCarrito.findIndex(
     (producto) => producto.id === idBoton
   );
-
   productosEnCarrito.splice(index, 1);
+  
   cargarProductosCarrito();
-
+  
   localStorage.setItem(
     "productps-en-carrito",
     JSON.stringify(productosEnCarrito)
+    
   );
 }
+
 
 botonVaciar.addEventListener("click", vaciarCarrito);
 
 function vaciarCarrito() {
-  productosEnCarrito.length = 0;
+  
+
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+  })
+swalWithBootstrapButtons.fire({
+  title: 'Queres eliminar el carrito completo?',
+  text: "No vas a poder revertir esto",
+  icon: 'warning',
+  showCancelButton: true,
+  confirmButtonText: 'Si, vaciar el carrito',
+  cancelButtonText: 'No, cancelalo!',
+  reverseButtons: true
+}).then((result) => {
+  if (result.isConfirmed) {
+    swalWithBootstrapButtons.fire(
+      'Eliminado!',
+      'El carrito fue eliminado correctamente',
+      'success'
+    )
+    productosEnCarrito.length = 0;
   localStorage.setItem(
     "productos-en-carrito",
     JSON.stringify(productosEnCarrito)
   );
   cargarProductosCarrito();
+  } else if (
+    /* Read more about handling dismissals below */
+    result.dismiss === Swal.DismissReason.cancel
+  ) {
+    swalWithBootstrapButtons.fire(
+      'Cancelado',
+      'Tu carrito esta seguro :)',
+      'error'
+    )
+  }
+})
+
+
+  
 }
 
 function actualizarTotal() {
@@ -103,6 +160,28 @@ function actualizarTotal() {
 botonComprar.addEventListener("click", comprarCarrito);
 
 function comprarCarrito() {
+  let timerInterval
+  Swal.fire({
+  title: 'Finalizando tu compra!',
+  html: 'Esta ventana se cerrara en  <b></b> milisegundos.',
+  timer: 3000,
+  timerProgressBar: true,
+  didOpen: () => {
+    Swal.showLoading()
+    const b = Swal.getHtmlContainer().querySelector('b')
+    timerInterval = setInterval(() => {
+      b.textContent = Swal.getTimerLeft()
+    }, 100)
+  },
+  willClose: () => {
+    clearInterval(timerInterval)
+  }
+}).then((result) => {
+  /* Read more about handling dismissals below */
+  if (result.dismiss === Swal.DismissReason.timer) {
+    console.log('I was closed by the timer')
+  }
+})
   productosEnCarrito.length = 0;
   localStorage.setItem(
     "productos-en-carrito",
